requirements.txt
Pillow
pyheif
pillow-heif
